fixed thread problem

+++ core/src/com/mygdx/crazybomber/server/ServerMap.java
@@ -20,8 +20,8 @@ public class ServerMap{
         _activeItemArray = new ArrayList<Item>();
         byte itemIDCounter = 0;
 
-        for (int i = 0; i < 15; i++) {
-            for (int j = 0; j < 15; j++) {
+        for (byte i = 0; i < 15; i++) {
+            for (byte j = 0; j < 15; j++) {
                 if (intMap[i][j] == 1) {
                     blockMatrix[i][j] = new UnbreakableBlock(i, j);
                 } else if (intMap[i][j] == 2) {
